Scope Elixir REPL buffers to projects
This change, in essence, allows a user to have multiple Elixir REPL
buffers with inf-elixir.

The general structure is that running `inf-elixir-project` will create
a project-specific REPL and `inf-elixir` will create a global
REPL.

If a REPL is already running for a given project When
`inf-elixir-project` is invoked, the user will be prompted as to
whether they want to keep the existing REPL or replace it.