Overhaul inf-elixir-send-* buffer lookup
This is a relatively complex flow that I've wanted to revisit for some
time. I will be uploading a diagram to the wiki that encapsulates most
of the flow.

Also introduces a new `inf-elixir-set-repl` command which allows the
user to manually set what REPL they want the current buffer to send
to. This choice is remembered until the buffer is killed or the
command is run again.